refactor: restructure fraud-detector into modular architecture with light/deep scan tiers
Reorganized fraud-detector from single file into modular structure:

**New architecture:**
- `index.ts` - Public API exports (scanLight, scanDeep, types)
- `core.ts` - FraudDetector class with pattern matching and analysis
- `extractor.ts` - Script extraction logic (inline, external, event handlers, JS URLs)
- `patterns.ts` - Weighted suspicious pattern definitions
- `types.ts` - All TypeScript interfaces and
